refactor(release): update command imports to use utils folder
Update imports in add-chart, remove-chart, update-qa, and update-unrc to use utils/errors.ts and utils/validation.ts.
